:bug: fix some bugs

+++ diff-highlight/install.fish
@@ -1,5 +1,9 @@
 #!/usr/bin/env fish
 
+if test -e /usr/local/bin/diff-highlight
+    exit 0
+end
+
 switch (uname)
 case Darwin
     sudo ln -s /usr/local/share/git-core/contrib/diff-highlight/diff-highlight /usr/local/bin/diff-highlight
